+++ common/src/main/java/earth/terrarium/adastra/common/container/VehicleContainer.java
@@ -5,6 +5,9 @@ import net.minecraft.nbt.ListTag;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.item.ItemStack;
 
+/**
+ * Simple container that retains the order stacks were inserted in.
+ */
 public class VehicleContainer extends SimpleContainer {
 
     public VehicleContainer(int size) {